feat: implement WhatsApp sharing functionality and modularize Markdown generation
